requirements.txt
hamolicious-vector-class==1.2.0
pygame==2.1.2
pygame-widgets==1.0.0
